requirements.txt
uvicorn
starlette~=0.37.2
python-dotenv~=1.0.1
python-multipart
requests~=2.32.3
fastapi~=0.111.0
pydantic~=2.7.3
virtualenv
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.21.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.36
ultralytics-thop==2.0.12
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.21.0